[TEST] - splitting unit tests out from GeoContextIT.  creating new tests for DataProviderConnection
